Update Error codes and refactor
