+++ user-profile/src/main/java/com/poc/moneymatter/controllers/UserController.java
@@ -5,7 +5,6 @@ import com.poc.moneymatter.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
-import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -32,12 +31,13 @@ public class UserController {
 
     @PostMapping(value = "/user")
     @ResponseBody
+    @ResponseStatus(HttpStatus.CREATED)
     public User save(@RequestBody User user) {
         return service.save(user);
     }
 
     @DeleteMapping(value = "/user/{id}")
-    @ResponseStatus(HttpStatus.OK)
+    @ResponseStatus(HttpStatus.NO_CONTENT)
     public void remove(@PathVariable("id") String id) {
         service.delete(UUID.fromString(id));
     }
@@ -45,7 +45,6 @@ public class UserController {
     @PutMapping(value = "/user")
     @ResponseBody
     public User update(@RequestBody User user) {
-        if (StringUtils.isEmpty(user.getId())) throw new IllegalArgumentException("User doesn't exist !");
         return service.save(user);
     }
 }
