+++ user-profile/src/main/java/com/poc/moneymatter/exceptions/MoneyMatterUserException.java
@@ -1,0 +1,9 @@
+package com.poc.moneymatter.exceptions;
+
+public class MoneyMatterUserException extends RuntimeException {
+
+    public MoneyMatterUserException(String exception) {
+        super(exception);
+    }
+
+}
