+++ user-profile/src/main/java/com/poc/moneymatter/exceptions/models/MoneyMatterErrorResponse.java
@@ -1,7 +1,9 @@
-package com.poc.moneymatter.exceptions;
+package com.poc.moneymatter.exceptions.models;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+
+import java.util.Date;
 
 @Data
 @AllArgsConstructor
@@ -11,5 +13,6 @@ public class MoneyMatterErrorResponse {
     private String cause;
     private String message;
     private Integer status;
+    private Date timestamp;
 
 }
