+++ user-profile/src/main/java/com/poc/moneymatter/services/UserService.java
@@ -2,8 +2,10 @@ package com.poc.moneymatter.services;
 
 import com.poc.moneymatter.dao.entity.User;
 import com.poc.moneymatter.dao.repository.UserRepository;
+import com.poc.moneymatter.exceptions.MoneyMatterUserException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 
 import java.util.List;
 import java.util.UUID;
@@ -15,6 +17,7 @@ public class UserService {
     private UserRepository repository;
 
     public User save(User user) {
+        if (StringUtils.isEmpty(user.getId())) throw new MoneyMatterUserException("Pease provide User ID !");
         return repository.save(user);
     }
